# improve performance of deserializing invalid json policies